+++ styles.js
@@ -1,4 +1,22 @@
 import { createGlobalStyle } from "styled-components";
+import { Roboto_Flex, Poppins } from "next/font/google";
+
+const roboto = Roboto_Flex({
+  subsets: [
+    "cyrillic",
+    "cyrillic-ext",
+    "greek",
+    "latin",
+    "latin-ext",
+    "vietnamese",
+  ],
+  weight: ["100", "200", "300", "400", "500", "600", "700", "800", "900"],
+});
+
+const poppins = Poppins({
+  subsets: ["latin"],
+  weight: ["100", "200", "300", "400", "500", "600", "700", "800", "900"],
+});
 
 export default createGlobalStyle`
   *,
@@ -9,6 +27,7 @@ export default createGlobalStyle`
   body {
     margin: 0;
     padding: 0;
+    font-family: ${roboto.style.fontFamily}, ${poppins.style.fontFamily}, Arial, sans-serif;
   }
 
   h2 {
